file access control uses a resource object for the current state
This change is meant to enable better why-run and reporting for file
based resources. FAC now takes a resource instead of the path to the
file in question. This enables why-run mode to simulate updates to
system state in the case where the file doesn't exist (since there's no
longer any need to call stat() on the file).

Instead of inspecting a file's ACL state right before changing it,
ScanAccessControl should be used to inspect the system's state and store
it in the current_resource object